backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic[email]==2.10.5
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
certifi==2023.11.17
PyPDF2==3.0.1
pdfplumber==0.9.0
# AI & ML Dependencies
groq==0.4.2
httpx==0.27.0
httpcore==1.0.5
numpy>=1.26.4
scikit-learn>=1.5.2
sentence-transformers>=3.3.1
torch>=2.5.1
# XGBoost Pipeline Dependencies
xgboost>=2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
joblib>=1.3.0
# Security
slowapi>=0.1.9
